+++ test/test_helper.rb
@@ -1,6 +1,7 @@
 require 'minitest/autorun'
 require 'minitest/spec'
 require 'active_support/core_ext/kernel/reporting'
+require 'active_model'
 
 $VERBOSE=true
 
@@ -10,6 +11,15 @@ module SimpleForm
       ActiveSupport::StringInquirer.new("test")
     end
   end
+end
+
+module Formtastic
+  module Rails
+    VERSION = ActiveSupport::VERSION
+  end
+end
+
+module RailsAdmin
 end
 
 require 'enumerize'